requirements.txt
paddleocr
mysql-connector
paddlepaddle
flask